refactor: move GHCi runtime behind facade
