+++ packages/agent-core/agent-core.cabal
@@ -65,6 +65,7 @@ library
                     , Agent.Tools.Grok.Terminal
                     , Agent.Tools.Ghci
                     , Agent.Tools.Ghci.Classify
+                    , Agent.Tools.Ghci.Runtime
                     , Agent.Tools.IO
                     , Agent.Tools.MultiAgents
                     , Agent.Tools.PlanMode
